refactor: split insightAPI.py into stats/snapshot/analytics modules
api/insightAPI.py was the single largest file in the repo at 1914 lines, with zero
test coverage: 4 routes, 3 of them small and independent, plus /api/insights - one
~1820-line route handler with ~30 nested closures computing the full analytics
payload (event enrichment, title/id mapping, lane computation, provider counts,
watchtime estimation, snapshot info).

Wrote characterization tests first (tests/test_insight_api.py, 8 tests) covering all
4 routes against the unmodified file, including two payload-shape tests for
/api/insights (empty-state defaults, and a populated fixture exercising history
breakdown/lanes/provider counts/watchtime) with a monkeypatched clock for
determinism. Confirmed all 8 pass against the original file before touching it.

Then split by responsibility, each a verbatim relocation - no logic changes:
- api/_insight_env.py: the shared _env() runtime lookup, extracted so the 3 route
  modules below can import it without a circular dependency on insightAPI.py
- api/insight_stats.py: /api/stats/raw, /api/stats
- api/insight_snapshot.py: /api/crosswatch/select-snapshot
- api/insight_analytics.py: /api/insights - the entire original route body, moved
  as one contiguous block (same nesting depth, so no reindentation needed either)
- api/insightAPI.py: now just wires the three register_insight_* functions together,
  keeping its existing `register_insights` import path (api/__init__.py:16 needed no
  changes)

Verified the move introduced zero drift: extracted each route body via `sed` (not
retyped) and diffed byte-for-byte against the original committed file - all three
match exactly. Also checked api/maintenanceAPI.py's cache-clear introspection over
`vars(insightAPI)` (matches names containing both an "insight"/"series" token and a
"cache"/"memo"/"state" token) - confirmed empty before and after the split, so no
change to that already-dormant maintenance hook.

Updated the characterization tests' monkeypatch targets (_env, time) to point at the
new module locations - the only test changes, mechanically following the code they
patch. Full suite passes with the same 3 pre-existing unrelated failures as main.
